# backend/docs/openapi.apifox.cn.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "供 APIFOX 导入的 OpenAPI 3.0 JSON。文档中的路径与 HTTP 方法集合与后端实际注册的 Gin 路由保持一致。"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地后端默认地址"
    }
  ],
...
  "paths": {
...
    "/sites/{siteID}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SiteIDParam"
        }
      ],
      "get": {
        "tags": ["站点"],
        "operationId": "downloadSiteRoot",
        "summary": "读取站点根文档",
        "description": "公开返回站点配置的索引文档。站点或对象查找失败时返回仅含状态码的网站响应。",
        "responses": {
          "200": {
            "description": "站点内容。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "站点标识无效。无响应体。"
          },
          "404": {
            "description": "站点或请求的内容不存在。若已配置错误文档，可能使用此状态码返回文档内容。"
          }
        }
      },
      "head": {
        "tags": ["站点"],
        "operationId": "headSiteRoot",
        "summary": "检查站点根文档",
        "description": "返回与 GET 相同的状态和表示头，但不返回响应体。",
        "responses": {
          "200": {
            "description": "站点内容元数据。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              }
            }
          },
          "400": {
            "description": "站点标识无效。无响应体。"
          },
          "404": {
            "description": "站点或请求的内容不存在。无响应体。"
          }
        }
      }
    },
    "/sites/{siteID}/{path}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SiteIDParam"
        },
        {
          "$ref": "#/components/parameters/SitePathParam"
        }
      ],
      "get": {
        "tags": ["站点"],
        "operationId": "downloadSitePath",
        "summary": "读取站点路径",
        "description": "公开返回站点根前缀下的文件。目录形式路径会解析到已配置的索引文档；SPA 回退和错误文档行为遵循站点配置。",
        "responses": {
          "200": {
            "description": "站点内容。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "站点标识或内容路径无效。无响应体。"
          },
          "404": {
            "description": "站点或请求的内容不存在。若已配置回退或错误文档，可能使用此状态码返回文档内容。"
          }
        }
      },
      "head": {
        "tags": ["站点"],
        "operationId": "headSitePath",
        "summary": "检查站点路径",
        "description": "返回与 GET 相同的状态和表示头，但不返回响应体。",
        "responses": {
          "200": {
            "description": "站点内容元数据。",
            "headers": {
              "Content-Type": {
                "$ref": "#/components/headers/ContentTypeHeader"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLengthHeader"
              },
              "ETag": {
                "$ref": "#/components/headers/ETagHeader"
              }
            }
          },
          "400": {
            "description": "站点标识或内容路径无效。无响应体。"
          },
          "404": {
            "description": "站点或请求的内容不存在。无响应体。"
          }
        }
      }
    },
...
    "/api/v1/sites": {
      "get": {
        "tags": ["站点"],
        "operationId": "listSites",
        "summary": "列出站点绑定",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "站点列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出站点。错误代码： site_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["站点"],
        "operationId": "createSite",
        "summary": "创建网站绑定",
        "description": "站点域名必须是合法 hostname。`root_prefix` 会被规范化为以斜杠结尾；`index_document` 在省略或为空时默认为 `index.html`；`enabled` 默认为 `true`；`spa_fallback` 默认为 `false`。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "站点已创建。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求正文或站点配置无效。错误代码： invalid_request、invalid_bucket_name、invalid_website_config、invalid_domain。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到存储桶。错误代码： bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "其中一个域已绑定到另一个站点。错误代码： domain_conflict。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "网站创建失败。错误代码： bucket_lookup_failed ， site_create_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sites/{siteID}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SiteIDParam"
        }
      ],
      "get": {
        "tags": ["站点"],
        "operationId": "getSite",
        "summary": "获取网站绑定",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "地点详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "siteID无效。错误代码： invalid_site_id。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "找不到站点。错误代码： site_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "查找站点失败。错误代码： site_lookup_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["站点"],
        "operationId": "updateSite",
        "summary": "更新网站绑定",
        "description": "与 `createSite` 相同的校验规则和默认值同样适用于传入的请求体。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "网站更新。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "siteID、请求正文或站点配置无效。错误代码： invalid_site_id、invalid_request、invalid_bucket_name、invalid_website_config、invalid_domain。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到站点或存储桶。错误代码： site_not_found, bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "其中一个域已绑定到另一个站点。错误代码： domain_conflict。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "站点更新失败。错误代码： bucket_lookup_failed ， site_update_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["站点"],
        "operationId": "deleteSite",
        "summary": "删除网站绑定",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "网站已删除。无响应正文。"
          },
          "400": {
            "description": "siteID无效。错误代码： invalid_site_id。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "找不到站点。错误代码： site_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "网站删除失败。错误代码： site_delete_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "SiteIDParam": {
        "name": "siteID",
        "in": "path",
        "required": true,
        "description": "正整数站点标识。",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
      "SitePathParam": {
        "name": "path",
        "in": "path",
        "required": true,
        "allowReserved": true,
        "style": "simple",
        "explode": false,
        "description": "站点相对内容路径。该占位符表示 Gin 的 /*path 通配符，可包含 '/'。传入不带开头斜杠的原始路径。",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
    },
    "headers": {
      "ContentTypeHeader": {
        "description": "存储的对象内容类型。",
        "schema": {
          "type": "string"
        }
      },
      "ContentLengthHeader": {
        "description": "存储的对象大小（以字节为单位）。",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 0
        }
      },
      "ETagHeader": {
        "description": "为存储的对象内容生成的实体标记。",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
...
      "SuccessEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": {
            "type": "string",
            "description": "中间件返回的请求关联ID。"
          },
          "data": {
            "type": "object",
            "description": "接口特定的响应数据。"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "SiteRequest": {
        "type": "object",
        "required": ["bucket"],
        "properties": {
          "bucket": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
            "minLength": 3,
            "maxLength": 63
          },
          "root_prefix": {
            "type": "string",
            "maxLength": 512,
            "default": "",
            "description": "根文件夹前缀。前导斜杠被去除。非空值被标准化为以“/”结尾。"
          },
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "index_document": {
            "type": "string",
            "maxLength": 512,
            "default": "index.html",
            "description": "相对于 `root_prefix` 的文档路径。默认为 `index.html`。"
          },
          "error_document": {
            "type": "string",
            "maxLength": 512,
            "default": "",
            "description": "相对于 `root_prefix` 的可选错误文档路径。"
          },
          "spa_fallback": {
            "type": "boolean",
            "default": false
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "可选的域名绑定。每个项目都必须是合法 hostname。"
          }
        }
      },
...
      "Site": {
        "type": "object",
        "required": [
          "id",
          "bucket",
          "root_prefix",
          "enabled",
          "index_document",
          "error_document",
          "spa_fallback",
          "domains",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "bucket": {
            "type": "string"
          },
          "root_prefix": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "index_document": {
            "type": "string"
          },
          "error_document": {
            "type": "string"
          },
          "spa_fallback": {
            "type": "boolean"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SiteList": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Site"
            }
          }
        }
      },
...
      "SiteEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/Site"
              }
            }
          }
        ]
      },
      "SiteListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/SiteList"
              }
            }
          }
        ]
      },
...
    }
  }
}